Moved header to a component
Closes #1153
